Fixed inventory bug, minor refactoring, start of abilities
Inventory Bug: items were no longer being "activated" when walking over them due to hierarchy change in straw manny.
Refactoring: cleaned up the inventory and pickup class a bit.
Abilities: started writing ability system that is setup to work alongside the pickups. When a ability is activated _use_ability is called on the base script that is meant to be overridden per child. This function takes in straw manny